pyproject.toml
[project]
name = "deboche-bot"
version = "0.1.0"
dependencies = [
  "discord.py",
  "python-dotenv",
  "cogwatch",
  "jishaku",
  "akinator.py",
  "beautifulsoup4",
  "playwright",
  "discord-ui",
  "yt-dlp",
  "PyNaCl",
  "google-genai",
  "aiosqlite"
]
